@@ -3,7 +3,7 @@ description: A new Flutter project.
 version: 1.2.2+13
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.7.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -24,31 +24,30 @@ dependencies:
   flutter_launcher_icons: ^0.7.4
   fluttertoast: ^3.1.3
   font_awesome_flutter: ^8.5.0
-  freezed_annotation: ^0.11.0
+  freezed_annotation: ^0.12.0
   geolocator: ^5.2.1
   get: ^3.24.0
   get_it: ^4.0.2
+  get_storage: ^1.4.0
   google_fonts: ^1.0.0
   image_downloader: ^0.19.1
   image_picker: ^0.6.1+10
-  injectable: ^0.4.0+1
+  injectable: ^1.0.7
   intl: ^0.16.1
   json_serializable: ^3.3.0
   material_design_icons_flutter: ^4.0.5145
   path: ^1.6.4
   path_provider: ^1.4.0
-  provider: ^4.0.5+1
   qr_flutter: ^3.0.1
   rflutter_alert: ^1.0.3
   share: ^0.6.3+5
-  shared_preferences: ^0.5.7+3
   url_launcher: ^5.2.5
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^1.10.0
-  freezed: ^0.11.0
-  injectable_generator: ^0.4.1
+  build_runner: ^1.10.13
+  freezed: ^0.12.7
+  injectable_generator: ^1.0.7
 
 flutter_icons:
   android: true
@@ -58,6 +57,9 @@ flutter_icons:
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 # The following section is specific to Flutter.
+dependency_overrides: null
+# analyzer: '0.39.17'
+# build_resolvers: '1.3.10'
 flutter:
 
   # The following line ensures that the Material Icons font is
