@@ -1,5 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named
+// ignore_for_file: deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
 
 part of 'transaction_bloc.dart';
 
@@ -9,9 +9,11 @@ part of 'transaction_bloc.dart';
 
 T _$identity<T>(T value) => value;
 
+/// @nodoc
 class _$TransactionEventTearOff {
   const _$TransactionEventTearOff();
 
+// ignore: unused_element
   _GetPointAndLastTrans getPointAndLastTrans(
       {@required GetPointAndLastTransRequest userId}) {
     return _GetPointAndLastTrans(
@@ -19,6 +21,7 @@ class _$TransactionEventTearOff {
     );
   }
 
+// ignore: unused_element
   _GetAllTransaction getAllTransaction({@required String userId}) {
     return _GetAllTransaction(
       userId: userId,
@@ -26,40 +29,44 @@ class _$TransactionEventTearOff {
   }
 }
 
+/// @nodoc
 // ignore: unused_element
 const $TransactionEvent = _$TransactionEventTearOff();
 
+/// @nodoc
 mixin _$TransactionEvent {
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result getPointAndLastTrans(GetPointAndLastTransRequest userId),
-    @required Result getAllTransaction(String userId),
+  TResult when<TResult extends Object>({
+    @required TResult getPointAndLastTrans(GetPointAndLastTransRequest userId),
+    @required TResult getAllTransaction(String userId),
   });
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result getPointAndLastTrans(GetPointAndLastTransRequest userId),
-    Result getAllTransaction(String userId),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object>({
+    TResult getPointAndLastTrans(GetPointAndLastTransRequest userId),
+    TResult getAllTransaction(String userId),
+    @required TResult orElse(),
   });
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result getPointAndLastTrans(_GetPointAndLastTrans value),
-    @required Result getAllTransaction(_GetAllTransaction value),
+  TResult map<TResult extends Object>({
+    @required TResult getPointAndLastTrans(_GetPointAndLastTrans value),
+    @required TResult getAllTransaction(_GetAllTransaction value),
   });
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result getPointAndLastTrans(_GetPointAndLastTrans value),
-    Result getAllTransaction(_GetAllTransaction value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object>({
+    TResult getPointAndLastTrans(_GetPointAndLastTrans value),
+    TResult getAllTransaction(_GetAllTransaction value),
+    @required TResult orElse(),
   });
 }
 
+/// @nodoc
 abstract class $TransactionEventCopyWith<$Res> {
   factory $TransactionEventCopyWith(
           TransactionEvent value, $Res Function(TransactionEvent) then) =
       _$TransactionEventCopyWithImpl<$Res>;
 }
 
+/// @nodoc
 class _$TransactionEventCopyWithImpl<$Res>
     implements $TransactionEventCopyWith<$Res> {
   _$TransactionEventCopyWithImpl(this._value, this._then);
@@ -69,6 +76,7 @@ class _$TransactionEventCopyWithImpl<$Res>
   final $Res Function(TransactionEvent) _then;
 }
 
+/// @nodoc
 abstract class _$GetPointAndLastTransCopyWith<$Res> {
   factory _$GetPointAndLastTransCopyWith(_GetPointAndLastTrans value,
           $Res Function(_GetPointAndLastTrans) then) =
@@ -76,6 +84,7 @@ abstract class _$GetPointAndLastTransCopyWith<$Res> {
   $Res call({GetPointAndLastTransRequest userId});
 }
 
+/// @nodoc
 class __$GetPointAndLastTransCopyWithImpl<$Res>
     extends _$TransactionEventCopyWithImpl<$Res>
     implements _$GetPointAndLastTransCopyWith<$Res> {
@@ -98,6 +107,7 @@ class __$GetPointAndLastTransCopyWithImpl<$Res>
   }
 }
 
+/// @nodoc
 class _$_GetPointAndLastTrans implements _GetPointAndLastTrans {
   _$_GetPointAndLastTrans({@required this.userId}) : assert(userId != null);
 
@@ -121,6 +131,7 @@ class _$_GetPointAndLastTrans implements _GetPointAndLastTrans {
   int get hashCode =>
       runtimeType.hashCode ^ const DeepCollectionEquality().hash(userId);
 
+  @JsonKey(ignore: true)
   @override
   _$GetPointAndLastTransCopyWith<_GetPointAndLastTrans> get copyWith =>
       __$GetPointAndLastTransCopyWithImpl<_GetPointAndLastTrans>(
@@ -128,9 +139,9 @@ class _$_GetPointAndLastTrans implements _GetPointAndLastTrans {
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result getPointAndLastTrans(GetPointAndLastTransRequest userId),
-    @required Result getAllTransaction(String userId),
+  TResult when<TResult extends Object>({
+    @required TResult getPointAndLastTrans(GetPointAndLastTransRequest userId),
+    @required TResult getAllTransaction(String userId),
   }) {
     assert(getPointAndLastTrans != null);
     assert(getAllTransaction != null);
@@ -139,10 +150,10 @@ class _$_GetPointAndLastTrans implements _GetPointAndLastTrans {
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result getPointAndLastTrans(GetPointAndLastTransRequest userId),
-    Result getAllTransaction(String userId),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object>({
+    TResult getPointAndLastTrans(GetPointAndLastTransRequest userId),
+    TResult getAllTransaction(String userId),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (getPointAndLastTrans != null) {
@@ -153,9 +164,9 @@ class _$_GetPointAndLastTrans implements _GetPointAndLastTrans {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result getPointAndLastTrans(_GetPointAndLastTrans value),
-    @required Result getAllTransaction(_GetAllTransaction value),
+  TResult map<TResult extends Object>({
+    @required TResult getPointAndLastTrans(_GetPointAndLastTrans value),
+    @required TResult getAllTransaction(_GetAllTransaction value),
   }) {
     assert(getPointAndLastTrans != null);
     assert(getAllTransaction != null);
@@ -164,10 +175,10 @@ class _$_GetPointAndLastTrans implements _GetPointAndLastTrans {
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result getPointAndLastTrans(_GetPointAndLastTrans value),
-    Result getAllTransaction(_GetAllTransaction value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object>({
+    TResult getPointAndLastTrans(_GetPointAndLastTrans value),
+    TResult getAllTransaction(_GetAllTransaction value),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (getPointAndLastTrans != null) {
@@ -182,9 +193,11 @@ abstract class _GetPointAndLastTrans implements TransactionEvent {
       {@required GetPointAndLastTransRequest userId}) = _$_GetPointAndLastTrans;
 
   GetPointAndLastTransRequest get userId;
+  @JsonKey(ignore: true)
   _$GetPointAndLastTransCopyWith<_GetPointAndLastTrans> get copyWith;
 }
 
+/// @nodoc
 abstract class _$GetAllTransactionCopyWith<$Res> {
   factory _$GetAllTransactionCopyWith(
           _GetAllTransaction value, $Res Function(_GetAllTransaction) then) =
@@ -192,6 +205,7 @@ abstract class _$GetAllTransactionCopyWith<$Res> {
   $Res call({String userId});
 }
 
+/// @nodoc
 class __$GetAllTransactionCopyWithImpl<$Res>
     extends _$TransactionEventCopyWithImpl<$Res>
     implements _$GetAllTransactionCopyWith<$Res> {
@@ -212,6 +226,7 @@ class __$GetAllTransactionCopyWithImpl<$Res>
   }
 }
 
+/// @nodoc
 class _$_GetAllTransaction implements _GetAllTransaction {
   _$_GetAllTransaction({@required this.userId}) : assert(userId != null);
 
@@ -235,15 +250,16 @@ class _$_GetAllTransaction implements _GetAllTransaction {
   int get hashCode =>
       runtimeType.hashCode ^ const DeepCollectionEquality().hash(userId);
 
+  @JsonKey(ignore: true)
   @override
   _$GetAllTransactionCopyWith<_GetAllTransaction> get copyWith =>
       __$GetAllTransactionCopyWithImpl<_GetAllTransaction>(this, _$identity);
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result getPointAndLastTrans(GetPointAndLastTransRequest userId),
-    @required Result getAllTransaction(String userId),
+  TResult when<TResult extends Object>({
+    @required TResult getPointAndLastTrans(GetPointAndLastTransRequest userId),
+    @required TResult getAllTransaction(String userId),
   }) {
     assert(getPointAndLastTrans != null);
     assert(getAllTransaction != null);
@@ -252,10 +268,10 @@ class _$_GetAllTransaction implements _GetAllTransaction {
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result getPointAndLastTrans(GetPointAndLastTransRequest userId),
-    Result getAllTransaction(String userId),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object>({
+    TResult getPointAndLastTrans(GetPointAndLastTransRequest userId),
+    TResult getAllTransaction(String userId),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (getAllTransaction != null) {
@@ -266,9 +282,9 @@ class _$_GetAllTransaction implements _GetAllTransaction {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result getPointAndLastTrans(_GetPointAndLastTrans value),
-    @required Result getAllTransaction(_GetAllTransaction value),
+  TResult map<TResult extends Object>({
+    @required TResult getPointAndLastTrans(_GetPointAndLastTrans value),
+    @required TResult getAllTransaction(_GetAllTransaction value),
   }) {
     assert(getPointAndLastTrans != null);
     assert(getAllTransaction != null);
@@ -277,10 +293,10 @@ class _$_GetAllTransaction implements _GetAllTransaction {
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result getPointAndLastTrans(_GetPointAndLastTrans value),
-    Result getAllTransaction(_GetAllTransaction value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object>({
+    TResult getPointAndLastTrans(_GetPointAndLastTrans value),
+    TResult getAllTransaction(_GetAllTransaction value),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (getAllTransaction != null) {
@@ -294,20 +310,25 @@ abstract class _GetAllTransaction implements TransactionEvent {
   factory _GetAllTransaction({@required String userId}) = _$_GetAllTransaction;
 
   String get userId;
+  @JsonKey(ignore: true)
   _$GetAllTransactionCopyWith<_GetAllTransaction> get copyWith;
 }
 
+/// @nodoc
 class _$TransactionStateTearOff {
   const _$TransactionStateTearOff();
 
+// ignore: unused_element
   _OnProgress onProgress() {
     return _OnProgress();
   }
 
+// ignore: unused_element
   _Initial initial() {
     return _Initial();
   }
 
+// ignore: unused_element
   _OnGetPointOption onGetPointOption(
       Option<Either<TransactionFailure, GetPointAndLastTransResponse>>
           failOrSuccessOption) {
@@ -317,49 +338,53 @@ class _$TransactionStateTearOff {
   }
 }
 
+/// @nodoc
 // ignore: unused_element
 const $TransactionState = _$TransactionStateTearOff();
 
+/// @nodoc
 mixin _$TransactionState {
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result onProgress(),
-    @required Result initial(),
+  TResult when<TResult extends Object>({
+    @required TResult onProgress(),
+    @required TResult initial(),
     @required
-        Result onGetPointOption(
+        TResult onGetPointOption(
             Option<Either<TransactionFailure, GetPointAndLastTransResponse>>
                 failOrSuccessOption),
   });
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result onProgress(),
-    Result initial(),
-    Result onGetPointOption(
+  TResult maybeWhen<TResult extends Object>({
+    TResult onProgress(),
+    TResult initial(),
+    TResult onGetPointOption(
         Option<Either<TransactionFailure, GetPointAndLastTransResponse>>
             failOrSuccessOption),
-    @required Result orElse(),
+    @required TResult orElse(),
   });
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result onProgress(_OnProgress value),
-    @required Result initial(_Initial value),
-    @required Result onGetPointOption(_OnGetPointOption value),
+  TResult map<TResult extends Object>({
+    @required TResult onProgress(_OnProgress value),
+    @required TResult initial(_Initial value),
+    @required TResult onGetPointOption(_OnGetPointOption value),
   });
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result onProgress(_OnProgress value),
-    Result initial(_Initial value),
-    Result onGetPointOption(_OnGetPointOption value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object>({
+    TResult onProgress(_OnProgress value),
+    TResult initial(_Initial value),
+    TResult onGetPointOption(_OnGetPointOption value),
+    @required TResult orElse(),
   });
 }
 
+/// @nodoc
 abstract class $TransactionStateCopyWith<$Res> {
   factory $TransactionStateCopyWith(
           TransactionState value, $Res Function(TransactionState) then) =
       _$TransactionStateCopyWithImpl<$Res>;
 }
 
+/// @nodoc
 class _$TransactionStateCopyWithImpl<$Res>
     implements $TransactionStateCopyWith<$Res> {
   _$TransactionStateCopyWithImpl(this._value, this._then);
@@ -369,12 +394,14 @@ class _$TransactionStateCopyWithImpl<$Res>
   final $Res Function(TransactionState) _then;
 }
 
+/// @nodoc
 abstract class _$OnProgressCopyWith<$Res> {
   factory _$OnProgressCopyWith(
           _OnProgress value, $Res Function(_OnProgress) then) =
       __$OnProgressCopyWithImpl<$Res>;
 }
 
+/// @nodoc
 class __$OnProgressCopyWithImpl<$Res>
     extends _$TransactionStateCopyWithImpl<$Res>
     implements _$OnProgressCopyWith<$Res> {
@@ -386,6 +413,7 @@ class __$OnProgressCopyWithImpl<$Res>
   _OnProgress get _value => super._value as _OnProgress;
 }
 
+/// @nodoc
 class _$_OnProgress implements _OnProgress {
   _$_OnProgress();
 
@@ -404,11 +432,11 @@ class _$_OnProgress implements _OnProgress {
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result onProgress(),
-    @required Result initial(),
+  TResult when<TResult extends Object>({
+    @required TResult onProgress(),
+    @required TResult initial(),
     @required
-        Result onGetPointOption(
+        TResult onGetPointOption(
             Option<Either<TransactionFailure, GetPointAndLastTransResponse>>
                 failOrSuccessOption),
   }) {
@@ -420,13 +448,13 @@ class _$_OnProgress implements _OnProgress {
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result onProgress(),
-    Result initial(),
-    Result onGetPointOption(
+  TResult maybeWhen<TResult extends Object>({
+    TResult onProgress(),
+    TResult initial(),
+    TResult onGetPointOption(
         Option<Either<TransactionFailure, GetPointAndLastTransResponse>>
             failOrSuccessOption),
-    @required Result orElse(),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (onProgress != null) {
@@ -437,10 +465,10 @@ class _$_OnProgress implements _OnProgress {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result onProgress(_OnProgress value),
-    @required Result initial(_Initial value),
-    @required Result onGetPointOption(_OnGetPointOption value),
+  TResult map<TResult extends Object>({
+    @required TResult onProgress(_OnProgress value),
+    @required TResult initial(_Initial value),
+    @required TResult onGetPointOption(_OnGetPointOption value),
   }) {
     assert(onProgress != null);
     assert(initial != null);
@@ -450,11 +478,11 @@ class _$_OnProgress implements _OnProgress {
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result onProgress(_OnProgress value),
-    Result initial(_Initial value),
-    Result onGetPointOption(_OnGetPointOption value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object>({
+    TResult onProgress(_OnProgress value),
+    TResult initial(_Initial value),
+    TResult onGetPointOption(_OnGetPointOption value),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (onProgress != null) {
@@ -468,11 +496,13 @@ abstract class _OnProgress implements TransactionState {
   factory _OnProgress() = _$_OnProgress;
 }
 
+/// @nodoc
 abstract class _$InitialCopyWith<$Res> {
   factory _$InitialCopyWith(_Initial value, $Res Function(_Initial) then) =
       __$InitialCopyWithImpl<$Res>;
 }
 
+/// @nodoc
 class __$InitialCopyWithImpl<$Res> extends _$TransactionStateCopyWithImpl<$Res>
     implements _$InitialCopyWith<$Res> {
   __$InitialCopyWithImpl(_Initial _value, $Res Function(_Initial) _then)
@@ -482,6 +512,7 @@ class __$InitialCopyWithImpl<$Res> extends _$TransactionStateCopyWithImpl<$Res>
   _Initial get _value => super._value as _Initial;
 }
 
+/// @nodoc
 class _$_Initial implements _Initial {
   _$_Initial();
 
@@ -500,11 +531,11 @@ class _$_Initial implements _Initial {
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result onProgress(),
-    @required Result initial(),
+  TResult when<TResult extends Object>({
+    @required TResult onProgress(),
+    @required TResult initial(),
     @required
-        Result onGetPointOption(
+        TResult onGetPointOption(
             Option<Either<TransactionFailure, GetPointAndLastTransResponse>>
                 failOrSuccessOption),
   }) {
@@ -516,13 +547,13 @@ class _$_Initial implements _Initial {
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result onProgress(),
-    Result initial(),
-    Result onGetPointOption(
+  TResult maybeWhen<TResult extends Object>({
+    TResult onProgress(),
+    TResult initial(),
+    TResult onGetPointOption(
         Option<Either<TransactionFailure, GetPointAndLastTransResponse>>
             failOrSuccessOption),
-    @required Result orElse(),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (initial != null) {
@@ -533,10 +564,10 @@ class _$_Initial implements _Initial {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result onProgress(_OnProgress value),
-    @required Result initial(_Initial value),
-    @required Result onGetPointOption(_OnGetPointOption value),
+  TResult map<TResult extends Object>({
+    @required TResult onProgress(_OnProgress value),
+    @required TResult initial(_Initial value),
+    @required TResult onGetPointOption(_OnGetPointOption value),
   }) {
     assert(onProgress != null);
     assert(initial != null);
@@ -546,11 +577,11 @@ class _$_Initial implements _Initial {
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result onProgress(_OnProgress value),
-    Result initial(_Initial value),
-    Result onGetPointOption(_OnGetPointOption value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object>({
+    TResult onProgress(_OnProgress value),
+    TResult initial(_Initial value),
+    TResult onGetPointOption(_OnGetPointOption value),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (initial != null) {
@@ -564,6 +595,7 @@ abstract class _Initial implements TransactionState {
   factory _Initial() = _$_Initial;
 }
 
+/// @nodoc
 abstract class _$OnGetPointOptionCopyWith<$Res> {
   factory _$OnGetPointOptionCopyWith(
           _OnGetPointOption value, $Res Function(_OnGetPointOption) then) =
@@ -573,6 +605,7 @@ abstract class _$OnGetPointOptionCopyWith<$Res> {
           failOrSuccessOption});
 }
 
+/// @nodoc
 class __$OnGetPointOptionCopyWithImpl<$Res>
     extends _$TransactionStateCopyWithImpl<$Res>
     implements _$OnGetPointOptionCopyWith<$Res> {
@@ -596,6 +629,7 @@ class __$OnGetPointOptionCopyWithImpl<$Res>
   }
 }
 
+/// @nodoc
 class _$_OnGetPointOption implements _OnGetPointOption {
   _$_OnGetPointOption(this.failOrSuccessOption)
       : assert(failOrSuccessOption != null);
@@ -623,17 +657,18 @@ class _$_OnGetPointOption implements _OnGetPointOption {
       runtimeType.hashCode ^
       const DeepCollectionEquality().hash(failOrSuccessOption);
 
+  @JsonKey(ignore: true)
   @override
   _$OnGetPointOptionCopyWith<_OnGetPointOption> get copyWith =>
       __$OnGetPointOptionCopyWithImpl<_OnGetPointOption>(this, _$identity);
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result onProgress(),
-    @required Result initial(),
+  TResult when<TResult extends Object>({
+    @required TResult onProgress(),
+    @required TResult initial(),
     @required
-        Result onGetPointOption(
+        TResult onGetPointOption(
             Option<Either<TransactionFailure, GetPointAndLastTransResponse>>
                 failOrSuccessOption),
   }) {
@@ -645,13 +680,13 @@ class _$_OnGetPointOption implements _OnGetPointOption {
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result onProgress(),
-    Result initial(),
-    Result onGetPointOption(
+  TResult maybeWhen<TResult extends Object>({
+    TResult onProgress(),
+    TResult initial(),
+    TResult onGetPointOption(
         Option<Either<TransactionFailure, GetPointAndLastTransResponse>>
             failOrSuccessOption),
-    @required Result orElse(),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (onGetPointOption != null) {
@@ -662,10 +697,10 @@ class _$_OnGetPointOption implements _OnGetPointOption {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result onProgress(_OnProgress value),
-    @required Result initial(_Initial value),
-    @required Result onGetPointOption(_OnGetPointOption value),
+  TResult map<TResult extends Object>({
+    @required TResult onProgress(_OnProgress value),
+    @required TResult initial(_Initial value),
+    @required TResult onGetPointOption(_OnGetPointOption value),
   }) {
     assert(onProgress != null);
     assert(initial != null);
@@ -675,11 +710,11 @@ class _$_OnGetPointOption implements _OnGetPointOption {
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result onProgress(_OnProgress value),
-    Result initial(_Initial value),
-    Result onGetPointOption(_OnGetPointOption value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object>({
+    TResult onProgress(_OnProgress value),
+    TResult initial(_Initial value),
+    TResult onGetPointOption(_OnGetPointOption value),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (onGetPointOption != null) {
@@ -696,5 +731,6 @@ abstract class _OnGetPointOption implements TransactionState {
 
   Option<Either<TransactionFailure, GetPointAndLastTransResponse>>
       get failOrSuccessOption;
+  @JsonKey(ignore: true)
   _$OnGetPointOptionCopyWith<_OnGetPointOption> get copyWith;
 }
