@@ -1,5 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named
+// ignore_for_file: deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
 
 part of 'auth_failure.dart';
 
@@ -9,70 +9,79 @@ part of 'auth_failure.dart';
 
 T _$identity<T>(T value) => value;
 
+/// @nodoc
 class _$AuthFailureTearOff {
   const _$AuthFailureTearOff();
 
+// ignore: unused_element
   _BadRequest badRequest({String errorMessage}) {
     return _BadRequest(
       errorMessage: errorMessage,
     );
   }
 
+// ignore: unused_element
   _ServerError serverError() {
     return const _ServerError();
   }
 
+// ignore: unused_element
   _NotFound notFound() {
     return const _NotFound();
   }
 
+// ignore: unused_element
   _DefaultError defaultError() {
     return const _DefaultError();
   }
 }
 
+/// @nodoc
 // ignore: unused_element
 const $AuthFailure = _$AuthFailureTearOff();
 
+/// @nodoc
 mixin _$AuthFailure {
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result badRequest(String errorMessage),
-    @required Result serverError(),
-    @required Result notFound(),
-    @required Result defaultError(),
+  TResult when<TResult extends Object>({
+    @required TResult badRequest(String errorMessage),
+    @required TResult serverError(),
+    @required TResult notFound(),
+    @required TResult defaultError(),
   });
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result badRequest(String errorMessage),
-    Result serverError(),
-    Result notFound(),
-    Result defaultError(),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object>({
+    TResult badRequest(String errorMessage),
+    TResult serverError(),
+    TResult notFound(),
+    TResult defaultError(),
+    @required TResult orElse(),
   });
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result badRequest(_BadRequest value),
-    @required Result serverError(_ServerError value),
-    @required Result notFound(_NotFound value),
-    @required Result defaultError(_DefaultError value),
+  TResult map<TResult extends Object>({
+    @required TResult badRequest(_BadRequest value),
+    @required TResult serverError(_ServerError value),
+    @required TResult notFound(_NotFound value),
+    @required TResult defaultError(_DefaultError value),
   });
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result badRequest(_BadRequest value),
-    Result serverError(_ServerError value),
-    Result notFound(_NotFound value),
-    Result defaultError(_DefaultError value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object>({
+    TResult badRequest(_BadRequest value),
+    TResult serverError(_ServerError value),
+    TResult notFound(_NotFound value),
+    TResult defaultError(_DefaultError value),
+    @required TResult orElse(),
   });
 }
 
+/// @nodoc
 abstract class $AuthFailureCopyWith<$Res> {
   factory $AuthFailureCopyWith(
           AuthFailure value, $Res Function(AuthFailure) then) =
       _$AuthFailureCopyWithImpl<$Res>;
 }
 
+/// @nodoc
 class _$AuthFailureCopyWithImpl<$Res> implements $AuthFailureCopyWith<$Res> {
   _$AuthFailureCopyWithImpl(this._value, this._then);
 
@@ -81,6 +90,7 @@ class _$AuthFailureCopyWithImpl<$Res> implements $AuthFailureCopyWith<$Res> {
   final $Res Function(AuthFailure) _then;
 }
 
+/// @nodoc
 abstract class _$BadRequestCopyWith<$Res> {
   factory _$BadRequestCopyWith(
           _BadRequest value, $Res Function(_BadRequest) then) =
@@ -88,6 +98,7 @@ abstract class _$BadRequestCopyWith<$Res> {
   $Res call({String errorMessage});
 }
 
+/// @nodoc
 class __$BadRequestCopyWithImpl<$Res> extends _$AuthFailureCopyWithImpl<$Res>
     implements _$BadRequestCopyWith<$Res> {
   __$BadRequestCopyWithImpl(
@@ -109,6 +120,7 @@ class __$BadRequestCopyWithImpl<$Res> extends _$AuthFailureCopyWithImpl<$Res>
   }
 }
 
+/// @nodoc
 class _$_BadRequest implements _BadRequest {
   const _$_BadRequest({this.errorMessage});
 
@@ -133,17 +145,18 @@ class _$_BadRequest implements _BadRequest {
   int get hashCode =>
       runtimeType.hashCode ^ const DeepCollectionEquality().hash(errorMessage);
 
+  @JsonKey(ignore: true)
   @override
   _$BadRequestCopyWith<_BadRequest> get copyWith =>
       __$BadRequestCopyWithImpl<_BadRequest>(this, _$identity);
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result badRequest(String errorMessage),
-    @required Result serverError(),
-    @required Result notFound(),
-    @required Result defaultError(),
+  TResult when<TResult extends Object>({
+    @required TResult badRequest(String errorMessage),
+    @required TResult serverError(),
+    @required TResult notFound(),
+    @required TResult defaultError(),
   }) {
     assert(badRequest != null);
     assert(serverError != null);
@@ -154,12 +167,12 @@ class _$_BadRequest implements _BadRequest {
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result badRequest(String errorMessage),
-    Result serverError(),
-    Result notFound(),
-    Result defaultError(),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object>({
+    TResult badRequest(String errorMessage),
+    TResult serverError(),
+    TResult notFound(),
+    TResult defaultError(),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (badRequest != null) {
@@ -170,11 +183,11 @@ class _$_BadRequest implements _BadRequest {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result badRequest(_BadRequest value),
-    @required Result serverError(_ServerError value),
-    @required Result notFound(_NotFound value),
-    @required Result defaultError(_DefaultError value),
+  TResult map<TResult extends Object>({
+    @required TResult badRequest(_BadRequest value),
+    @required TResult serverError(_ServerError value),
+    @required TResult notFound(_NotFound value),
+    @required TResult defaultError(_DefaultError value),
   }) {
     assert(badRequest != null);
     assert(serverError != null);
@@ -185,12 +198,12 @@ class _$_BadRequest implements _BadRequest {
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result badRequest(_BadRequest value),
-    Result serverError(_ServerError value),
-    Result notFound(_NotFound value),
-    Result defaultError(_DefaultError value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object>({
+    TResult badRequest(_BadRequest value),
+    TResult serverError(_ServerError value),
+    TResult notFound(_NotFound value),
+    TResult defaultError(_DefaultError value),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (badRequest != null) {
@@ -204,15 +217,18 @@ abstract class _BadRequest implements AuthFailure {
   const factory _BadRequest({String errorMessage}) = _$_BadRequest;
 
   String get errorMessage;
+  @JsonKey(ignore: true)
   _$BadRequestCopyWith<_BadRequest> get copyWith;
 }
 
+/// @nodoc
 abstract class _$ServerErrorCopyWith<$Res> {
   factory _$ServerErrorCopyWith(
           _ServerError value, $Res Function(_ServerError) then) =
       __$ServerErrorCopyWithImpl<$Res>;
 }
 
+/// @nodoc
 class __$ServerErrorCopyWithImpl<$Res> extends _$AuthFailureCopyWithImpl<$Res>
     implements _$ServerErrorCopyWith<$Res> {
   __$ServerErrorCopyWithImpl(
@@ -223,6 +239,7 @@ class __$ServerErrorCopyWithImpl<$Res> extends _$AuthFailureCopyWithImpl<$Res>
   _ServerError get _value => super._value as _ServerError;
 }
 
+/// @nodoc
 class _$_ServerError implements _ServerError {
   const _$_ServerError();
 
@@ -241,11 +258,11 @@ class _$_ServerError implements _ServerError {
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result badRequest(String errorMessage),
-    @required Result serverError(),
-    @required Result notFound(),
-    @required Result defaultError(),
+  TResult when<TResult extends Object>({
+    @required TResult badRequest(String errorMessage),
+    @required TResult serverError(),
+    @required TResult notFound(),
+    @required TResult defaultError(),
   }) {
     assert(badRequest != null);
     assert(serverError != null);
@@ -256,12 +273,12 @@ class _$_ServerError implements _ServerError {
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result badRequest(String errorMessage),
-    Result serverError(),
-    Result notFound(),
-    Result defaultError(),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object>({
+    TResult badRequest(String errorMessage),
+    TResult serverError(),
+    TResult notFound(),
+    TResult defaultError(),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (serverError != null) {
@@ -272,11 +289,11 @@ class _$_ServerError implements _ServerError {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result badRequest(_BadRequest value),
-    @required Result serverError(_ServerError value),
-    @required Result notFound(_NotFound value),
-    @required Result defaultError(_DefaultError value),
+  TResult map<TResult extends Object>({
+    @required TResult badRequest(_BadRequest value),
+    @required TResult serverError(_ServerError value),
+    @required TResult notFound(_NotFound value),
+    @required TResult defaultError(_DefaultError value),
   }) {
     assert(badRequest != null);
     assert(serverError != null);
@@ -287,12 +304,12 @@ class _$_ServerError implements _ServerError {
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result badRequest(_BadRequest value),
-    Result serverError(_ServerError value),
-    Result notFound(_NotFound value),
-    Result defaultError(_DefaultError value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object>({
+    TResult badRequest(_BadRequest value),
+    TResult serverError(_ServerError value),
+    TResult notFound(_NotFound value),
+    TResult defaultError(_DefaultError value),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (serverError != null) {
@@ -306,11 +323,13 @@ abstract class _ServerError implements AuthFailure {
   const factory _ServerError() = _$_ServerError;
 }
 
+/// @nodoc
 abstract class _$NotFoundCopyWith<$Res> {
   factory _$NotFoundCopyWith(_NotFound value, $Res Function(_NotFound) then) =
       __$NotFoundCopyWithImpl<$Res>;
 }
 
+/// @nodoc
 class __$NotFoundCopyWithImpl<$Res> extends _$AuthFailureCopyWithImpl<$Res>
     implements _$NotFoundCopyWith<$Res> {
   __$NotFoundCopyWithImpl(_NotFound _value, $Res Function(_NotFound) _then)
@@ -320,6 +339,7 @@ class __$NotFoundCopyWithImpl<$Res> extends _$AuthFailureCopyWithImpl<$Res>
   _NotFound get _value => super._value as _NotFound;
 }
 
+/// @nodoc
 class _$_NotFound implements _NotFound {
   const _$_NotFound();
 
@@ -338,11 +358,11 @@ class _$_NotFound implements _NotFound {
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result badRequest(String errorMessage),
-    @required Result serverError(),
-    @required Result notFound(),
-    @required Result defaultError(),
+  TResult when<TResult extends Object>({
+    @required TResult badRequest(String errorMessage),
+    @required TResult serverError(),
+    @required TResult notFound(),
+    @required TResult defaultError(),
   }) {
     assert(badRequest != null);
     assert(serverError != null);
@@ -353,12 +373,12 @@ class _$_NotFound implements _NotFound {
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result badRequest(String errorMessage),
-    Result serverError(),
-    Result notFound(),
-    Result defaultError(),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object>({
+    TResult badRequest(String errorMessage),
+    TResult serverError(),
+    TResult notFound(),
+    TResult defaultError(),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (notFound != null) {
@@ -369,11 +389,11 @@ class _$_NotFound implements _NotFound {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result badRequest(_BadRequest value),
-    @required Result serverError(_ServerError value),
-    @required Result notFound(_NotFound value),
-    @required Result defaultError(_DefaultError value),
+  TResult map<TResult extends Object>({
+    @required TResult badRequest(_BadRequest value),
+    @required TResult serverError(_ServerError value),
+    @required TResult notFound(_NotFound value),
+    @required TResult defaultError(_DefaultError value),
   }) {
     assert(badRequest != null);
     assert(serverError != null);
@@ -384,12 +404,12 @@ class _$_NotFound implements _NotFound {
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result badRequest(_BadRequest value),
-    Result serverError(_ServerError value),
-    Result notFound(_NotFound value),
-    Result defaultError(_DefaultError value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object>({
+    TResult badRequest(_BadRequest value),
+    TResult serverError(_ServerError value),
+    TResult notFound(_NotFound value),
+    TResult defaultError(_DefaultError value),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (notFound != null) {
@@ -403,12 +423,14 @@ abstract class _NotFound implements AuthFailure {
   const factory _NotFound() = _$_NotFound;
 }
 
+/// @nodoc
 abstract class _$DefaultErrorCopyWith<$Res> {
   factory _$DefaultErrorCopyWith(
           _DefaultError value, $Res Function(_DefaultError) then) =
       __$DefaultErrorCopyWithImpl<$Res>;
 }
 
+/// @nodoc
 class __$DefaultErrorCopyWithImpl<$Res> extends _$AuthFailureCopyWithImpl<$Res>
     implements _$DefaultErrorCopyWith<$Res> {
   __$DefaultErrorCopyWithImpl(
@@ -419,6 +441,7 @@ class __$DefaultErrorCopyWithImpl<$Res> extends _$AuthFailureCopyWithImpl<$Res>
   _DefaultError get _value => super._value as _DefaultError;
 }
 
+/// @nodoc
 class _$_DefaultError implements _DefaultError {
   const _$_DefaultError();
 
@@ -437,11 +460,11 @@ class _$_DefaultError implements _DefaultError {
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result badRequest(String errorMessage),
-    @required Result serverError(),
-    @required Result notFound(),
-    @required Result defaultError(),
+  TResult when<TResult extends Object>({
+    @required TResult badRequest(String errorMessage),
+    @required TResult serverError(),
+    @required TResult notFound(),
+    @required TResult defaultError(),
   }) {
     assert(badRequest != null);
     assert(serverError != null);
@@ -452,12 +475,12 @@ class _$_DefaultError implements _DefaultError {
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result badRequest(String errorMessage),
-    Result serverError(),
-    Result notFound(),
-    Result defaultError(),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object>({
+    TResult badRequest(String errorMessage),
+    TResult serverError(),
+    TResult notFound(),
+    TResult defaultError(),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (defaultError != null) {
@@ -468,11 +491,11 @@ class _$_DefaultError implements _DefaultError {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result badRequest(_BadRequest value),
-    @required Result serverError(_ServerError value),
-    @required Result notFound(_NotFound value),
-    @required Result defaultError(_DefaultError value),
+  TResult map<TResult extends Object>({
+    @required TResult badRequest(_BadRequest value),
+    @required TResult serverError(_ServerError value),
+    @required TResult notFound(_NotFound value),
+    @required TResult defaultError(_DefaultError value),
   }) {
     assert(badRequest != null);
     assert(serverError != null);
@@ -483,12 +506,12 @@ class _$_DefaultError implements _DefaultError {
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result badRequest(_BadRequest value),
-    Result serverError(_ServerError value),
-    Result notFound(_NotFound value),
-    Result defaultError(_DefaultError value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object>({
+    TResult badRequest(_BadRequest value),
+    TResult serverError(_ServerError value),
+    TResult notFound(_NotFound value),
+    TResult defaultError(_DefaultError value),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (defaultError != null) {
