@@ -1,5 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named
+// ignore_for_file: deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
 
 part of 'feed_failure.dart';
 
@@ -9,68 +9,77 @@ part of 'feed_failure.dart';
 
 T _$identity<T>(T value) => value;
 
+/// @nodoc
 class _$FeedFailureTearOff {
   const _$FeedFailureTearOff();
 
+// ignore: unused_element
   _BadRequest badRequest() {
     return const _BadRequest();
   }
 
+// ignore: unused_element
   _ServerError serverError() {
     return const _ServerError();
   }
 
+// ignore: unused_element
   _NotFound notFound() {
     return const _NotFound();
   }
 
+// ignore: unused_element
   _DataIsEmpty dataIsEmpty() {
     return const _DataIsEmpty();
   }
 }
 
+/// @nodoc
 // ignore: unused_element
 const $FeedFailure = _$FeedFailureTearOff();
 
+/// @nodoc
 mixin _$FeedFailure {
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result badRequest(),
-    @required Result serverError(),
-    @required Result notFound(),
-    @required Result dataIsEmpty(),
+  TResult when<TResult extends Object>({
+    @required TResult badRequest(),
+    @required TResult serverError(),
+    @required TResult notFound(),
+    @required TResult dataIsEmpty(),
   });
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result badRequest(),
-    Result serverError(),
-    Result notFound(),
-    Result dataIsEmpty(),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object>({
+    TResult badRequest(),
+    TResult serverError(),
+    TResult notFound(),
+    TResult dataIsEmpty(),
+    @required TResult orElse(),
   });
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result badRequest(_BadRequest value),
-    @required Result serverError(_ServerError value),
-    @required Result notFound(_NotFound value),
-    @required Result dataIsEmpty(_DataIsEmpty value),
+  TResult map<TResult extends Object>({
+    @required TResult badRequest(_BadRequest value),
+    @required TResult serverError(_ServerError value),
+    @required TResult notFound(_NotFound value),
+    @required TResult dataIsEmpty(_DataIsEmpty value),
   });
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result badRequest(_BadRequest value),
-    Result serverError(_ServerError value),
-    Result notFound(_NotFound value),
-    Result dataIsEmpty(_DataIsEmpty value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object>({
+    TResult badRequest(_BadRequest value),
+    TResult serverError(_ServerError value),
+    TResult notFound(_NotFound value),
+    TResult dataIsEmpty(_DataIsEmpty value),
+    @required TResult orElse(),
   });
 }
 
+/// @nodoc
 abstract class $FeedFailureCopyWith<$Res> {
   factory $FeedFailureCopyWith(
           FeedFailure value, $Res Function(FeedFailure) then) =
       _$FeedFailureCopyWithImpl<$Res>;
 }
 
+/// @nodoc
 class _$FeedFailureCopyWithImpl<$Res> implements $FeedFailureCopyWith<$Res> {
   _$FeedFailureCopyWithImpl(this._value, this._then);
 
@@ -79,12 +88,14 @@ class _$FeedFailureCopyWithImpl<$Res> implements $FeedFailureCopyWith<$Res> {
   final $Res Function(FeedFailure) _then;
 }
 
+/// @nodoc
 abstract class _$BadRequestCopyWith<$Res> {
   factory _$BadRequestCopyWith(
           _BadRequest value, $Res Function(_BadRequest) then) =
       __$BadRequestCopyWithImpl<$Res>;
 }
 
+/// @nodoc
 class __$BadRequestCopyWithImpl<$Res> extends _$FeedFailureCopyWithImpl<$Res>
     implements _$BadRequestCopyWith<$Res> {
   __$BadRequestCopyWithImpl(
@@ -95,6 +106,7 @@ class __$BadRequestCopyWithImpl<$Res> extends _$FeedFailureCopyWithImpl<$Res>
   _BadRequest get _value => super._value as _BadRequest;
 }
 
+/// @nodoc
 class _$_BadRequest implements _BadRequest {
   const _$_BadRequest();
 
@@ -113,11 +125,11 @@ class _$_BadRequest implements _BadRequest {
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result badRequest(),
-    @required Result serverError(),
-    @required Result notFound(),
-    @required Result dataIsEmpty(),
+  TResult when<TResult extends Object>({
+    @required TResult badRequest(),
+    @required TResult serverError(),
+    @required TResult notFound(),
+    @required TResult dataIsEmpty(),
   }) {
     assert(badRequest != null);
     assert(serverError != null);
@@ -128,12 +140,12 @@ class _$_BadRequest implements _BadRequest {
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result badRequest(),
-    Result serverError(),
-    Result notFound(),
-    Result dataIsEmpty(),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object>({
+    TResult badRequest(),
+    TResult serverError(),
+    TResult notFound(),
+    TResult dataIsEmpty(),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (badRequest != null) {
@@ -144,11 +156,11 @@ class _$_BadRequest implements _BadRequest {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result badRequest(_BadRequest value),
-    @required Result serverError(_ServerError value),
-    @required Result notFound(_NotFound value),
-    @required Result dataIsEmpty(_DataIsEmpty value),
+  TResult map<TResult extends Object>({
+    @required TResult badRequest(_BadRequest value),
+    @required TResult serverError(_ServerError value),
+    @required TResult notFound(_NotFound value),
+    @required TResult dataIsEmpty(_DataIsEmpty value),
   }) {
     assert(badRequest != null);
     assert(serverError != null);
@@ -159,12 +171,12 @@ class _$_BadRequest implements _BadRequest {
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result badRequest(_BadRequest value),
-    Result serverError(_ServerError value),
-    Result notFound(_NotFound value),
-    Result dataIsEmpty(_DataIsEmpty value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object>({
+    TResult badRequest(_BadRequest value),
+    TResult serverError(_ServerError value),
+    TResult notFound(_NotFound value),
+    TResult dataIsEmpty(_DataIsEmpty value),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (badRequest != null) {
@@ -178,12 +190,14 @@ abstract class _BadRequest implements FeedFailure {
   const factory _BadRequest() = _$_BadRequest;
 }
 
+/// @nodoc
 abstract class _$ServerErrorCopyWith<$Res> {
   factory _$ServerErrorCopyWith(
           _ServerError value, $Res Function(_ServerError) then) =
       __$ServerErrorCopyWithImpl<$Res>;
 }
 
+/// @nodoc
 class __$ServerErrorCopyWithImpl<$Res> extends _$FeedFailureCopyWithImpl<$Res>
     implements _$ServerErrorCopyWith<$Res> {
   __$ServerErrorCopyWithImpl(
@@ -194,6 +208,7 @@ class __$ServerErrorCopyWithImpl<$Res> extends _$FeedFailureCopyWithImpl<$Res>
   _ServerError get _value => super._value as _ServerError;
 }
 
+/// @nodoc
 class _$_ServerError implements _ServerError {
   const _$_ServerError();
 
@@ -212,11 +227,11 @@ class _$_ServerError implements _ServerError {
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result badRequest(),
-    @required Result serverError(),
-    @required Result notFound(),
-    @required Result dataIsEmpty(),
+  TResult when<TResult extends Object>({
+    @required TResult badRequest(),
+    @required TResult serverError(),
+    @required TResult notFound(),
+    @required TResult dataIsEmpty(),
   }) {
     assert(badRequest != null);
     assert(serverError != null);
@@ -227,12 +242,12 @@ class _$_ServerError implements _ServerError {
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result badRequest(),
-    Result serverError(),
-    Result notFound(),
-    Result dataIsEmpty(),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object>({
+    TResult badRequest(),
+    TResult serverError(),
+    TResult notFound(),
+    TResult dataIsEmpty(),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (serverError != null) {
@@ -243,11 +258,11 @@ class _$_ServerError implements _ServerError {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result badRequest(_BadRequest value),
-    @required Result serverError(_ServerError value),
-    @required Result notFound(_NotFound value),
-    @required Result dataIsEmpty(_DataIsEmpty value),
+  TResult map<TResult extends Object>({
+    @required TResult badRequest(_BadRequest value),
+    @required TResult serverError(_ServerError value),
+    @required TResult notFound(_NotFound value),
+    @required TResult dataIsEmpty(_DataIsEmpty value),
   }) {
     assert(badRequest != null);
     assert(serverError != null);
@@ -258,12 +273,12 @@ class _$_ServerError implements _ServerError {
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result badRequest(_BadRequest value),
-    Result serverError(_ServerError value),
-    Result notFound(_NotFound value),
-    Result dataIsEmpty(_DataIsEmpty value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object>({
+    TResult badRequest(_BadRequest value),
+    TResult serverError(_ServerError value),
+    TResult notFound(_NotFound value),
+    TResult dataIsEmpty(_DataIsEmpty value),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (serverError != null) {
@@ -277,11 +292,13 @@ abstract class _ServerError implements FeedFailure {
   const factory _ServerError() = _$_ServerError;
 }
 
+/// @nodoc
 abstract class _$NotFoundCopyWith<$Res> {
   factory _$NotFoundCopyWith(_NotFound value, $Res Function(_NotFound) then) =
       __$NotFoundCopyWithImpl<$Res>;
 }
 
+/// @nodoc
 class __$NotFoundCopyWithImpl<$Res> extends _$FeedFailureCopyWithImpl<$Res>
     implements _$NotFoundCopyWith<$Res> {
   __$NotFoundCopyWithImpl(_NotFound _value, $Res Function(_NotFound) _then)
@@ -291,6 +308,7 @@ class __$NotFoundCopyWithImpl<$Res> extends _$FeedFailureCopyWithImpl<$Res>
   _NotFound get _value => super._value as _NotFound;
 }
 
+/// @nodoc
 class _$_NotFound implements _NotFound {
   const _$_NotFound();
 
@@ -309,11 +327,11 @@ class _$_NotFound implements _NotFound {
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result badRequest(),
-    @required Result serverError(),
-    @required Result notFound(),
-    @required Result dataIsEmpty(),
+  TResult when<TResult extends Object>({
+    @required TResult badRequest(),
+    @required TResult serverError(),
+    @required TResult notFound(),
+    @required TResult dataIsEmpty(),
   }) {
     assert(badRequest != null);
     assert(serverError != null);
@@ -324,12 +342,12 @@ class _$_NotFound implements _NotFound {
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result badRequest(),
-    Result serverError(),
-    Result notFound(),
-    Result dataIsEmpty(),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object>({
+    TResult badRequest(),
+    TResult serverError(),
+    TResult notFound(),
+    TResult dataIsEmpty(),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (notFound != null) {
@@ -340,11 +358,11 @@ class _$_NotFound implements _NotFound {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result badRequest(_BadRequest value),
-    @required Result serverError(_ServerError value),
-    @required Result notFound(_NotFound value),
-    @required Result dataIsEmpty(_DataIsEmpty value),
+  TResult map<TResult extends Object>({
+    @required TResult badRequest(_BadRequest value),
+    @required TResult serverError(_ServerError value),
+    @required TResult notFound(_NotFound value),
+    @required TResult dataIsEmpty(_DataIsEmpty value),
   }) {
     assert(badRequest != null);
     assert(serverError != null);
@@ -355,12 +373,12 @@ class _$_NotFound implements _NotFound {
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result badRequest(_BadRequest value),
-    Result serverError(_ServerError value),
-    Result notFound(_NotFound value),
-    Result dataIsEmpty(_DataIsEmpty value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object>({
+    TResult badRequest(_BadRequest value),
+    TResult serverError(_ServerError value),
+    TResult notFound(_NotFound value),
+    TResult dataIsEmpty(_DataIsEmpty value),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (notFound != null) {
@@ -374,12 +392,14 @@ abstract class _NotFound implements FeedFailure {
   const factory _NotFound() = _$_NotFound;
 }
 
+/// @nodoc
 abstract class _$DataIsEmptyCopyWith<$Res> {
   factory _$DataIsEmptyCopyWith(
           _DataIsEmpty value, $Res Function(_DataIsEmpty) then) =
       __$DataIsEmptyCopyWithImpl<$Res>;
 }
 
+/// @nodoc
 class __$DataIsEmptyCopyWithImpl<$Res> extends _$FeedFailureCopyWithImpl<$Res>
     implements _$DataIsEmptyCopyWith<$Res> {
   __$DataIsEmptyCopyWithImpl(
@@ -390,6 +410,7 @@ class __$DataIsEmptyCopyWithImpl<$Res> extends _$FeedFailureCopyWithImpl<$Res>
   _DataIsEmpty get _value => super._value as _DataIsEmpty;
 }
 
+/// @nodoc
 class _$_DataIsEmpty implements _DataIsEmpty {
   const _$_DataIsEmpty();
 
@@ -408,11 +429,11 @@ class _$_DataIsEmpty implements _DataIsEmpty {
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result badRequest(),
-    @required Result serverError(),
-    @required Result notFound(),
-    @required Result dataIsEmpty(),
+  TResult when<TResult extends Object>({
+    @required TResult badRequest(),
+    @required TResult serverError(),
+    @required TResult notFound(),
+    @required TResult dataIsEmpty(),
   }) {
     assert(badRequest != null);
     assert(serverError != null);
@@ -423,12 +444,12 @@ class _$_DataIsEmpty implements _DataIsEmpty {
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result badRequest(),
-    Result serverError(),
-    Result notFound(),
-    Result dataIsEmpty(),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object>({
+    TResult badRequest(),
+    TResult serverError(),
+    TResult notFound(),
+    TResult dataIsEmpty(),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (dataIsEmpty != null) {
@@ -439,11 +460,11 @@ class _$_DataIsEmpty implements _DataIsEmpty {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result badRequest(_BadRequest value),
-    @required Result serverError(_ServerError value),
-    @required Result notFound(_NotFound value),
-    @required Result dataIsEmpty(_DataIsEmpty value),
+  TResult map<TResult extends Object>({
+    @required TResult badRequest(_BadRequest value),
+    @required TResult serverError(_ServerError value),
+    @required TResult notFound(_NotFound value),
+    @required TResult dataIsEmpty(_DataIsEmpty value),
   }) {
     assert(badRequest != null);
     assert(serverError != null);
@@ -454,12 +475,12 @@ class _$_DataIsEmpty implements _DataIsEmpty {
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result badRequest(_BadRequest value),
-    Result serverError(_ServerError value),
-    Result notFound(_NotFound value),
-    Result dataIsEmpty(_DataIsEmpty value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object>({
+    TResult badRequest(_BadRequest value),
+    TResult serverError(_ServerError value),
+    TResult notFound(_NotFound value),
+    TResult dataIsEmpty(_DataIsEmpty value),
+    @required TResult orElse(),
   }) {
     assert(orElse != null);
     if (dataIsEmpty != null) {
