@@ -1,7 +1,6 @@
 import 'package:get_it/get_it.dart';
+import 'package:i_love_iruka/injection.config.dart';
 import 'package:injectable/injectable.dart';
-
-import 'injection.iconfig.dart';
 
 final GetIt getIt = GetIt.instance;
 
