@@ -3,13 +3,15 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_datetime_picker/flutter_datetime_picker.dart';
 import 'package:fluttertoast/fluttertoast.dart';
+import 'package:get/get.dart';
 import 'package:i_love_iruka/application/auth/auth_bloc.dart';
-import 'package:i_love_iruka/application/auth/auth_provider.dart';
+import 'package:i_love_iruka/application/auth/auth_controller.dart';
 import 'package:i_love_iruka/application/auth/authentication/authentication_bloc.dart';
 import 'package:i_love_iruka/application/other/other_bloc.dart';
 import 'package:i_love_iruka/domain/core/user.dart';
 import 'package:i_love_iruka/domain/core/value_actions.dart';
 import 'package:i_love_iruka/domain/other/raja_ongkir.dart';
+import 'package:i_love_iruka/infrastructure/core/local_storage.dart';
 import 'package:i_love_iruka/injection.dart';
 import 'package:i_love_iruka/presentation/widgets/btn_primarary_blue_loading.dart';
 import 'package:i_love_iruka/presentation/widgets/btn_primary_blue.dart';
@@ -70,13 +72,16 @@ class _AdvancedInformationContentState
     Capability(value: 3, name: "Expert"),
   ];
 
+  final provinceBloc = getIt<OtherBloc>();
+  final cityBloc = getIt<OtherBloc>();
   User _currentUser;
 
   @override
   void initState() {
     isFirstChanged = false;
     //get data from provider and move
-    _currentUser = context.read<AuthProvider>().getUserData;
+
+    _currentUser = Get.put(AuthController()).getUserData();
     //fill controller from provider and user data
     _expYearsController =
         TextEditingController(text: _currentUser.yearsOfExperience.toString());
@@ -137,8 +142,8 @@ class _AdvancedInformationContentState
 
   @override
   Widget build(BuildContext context) {
-    return Consumer<AuthProvider>(
-      builder: (context, authProvider, _) => MultiBlocProvider(
+    return GetBuilder<AuthController>(
+      builder: (_auth) => MultiBlocProvider(
         providers: [
           BlocProvider(
               create: (context) =>
@@ -146,101 +151,65 @@ class _AdvancedInformationContentState
           BlocProvider(create: (context) => getIt<AuthBloc>()),
         ],
         child: BlocListener<AuthBloc, AuthState>(
-          listener: (context, state) {
-            state.maybeMap(
-                orElse: () {},
-                onProgress: (e) {
-                  flushbar = showFlushbarLoading(loading: "Updating Account...")
-                    ..show(context);
-                },
-                failOrSuccessUpdateCustomerOption: (e) {
-                  dismissFlushbar(flushbar);
-                  e.updateCustomerOption.fold(
-                      () => null,
+            listener: (context, state) {
+              state.maybeMap(
+                  orElse: () {},
+                  onProgress: (e) {
+                    flushbar =
+                        showFlushbarLoading(loading: "Updating Account...")
+                          ..show(context);
+                  },
+                  failOrSuccessUpdateCustomerOption: (e) {
+                    dismissFlushbar(flushbar);
+                    e.updateCustomerOption.fold(
+                        () => null,
+                        (a) => a.fold(
+                              (l) => showFlushbarError(
+                                  errMessage: "Can not update data")
+                                ..show(context),
+                              (r) {
+                                showFlushbarSuccess(
+                                    succMessage: "Data Updated Successfull")
+                                  ..show(context);
+                                r.map(
+                                    loginRequestData: (e) {},
+                                    loginResponseData: (resp) {
+                                      _auth.setUserData(resp.user);
+
+                                      saveUserData(resp.user);
+                                    });
+                              },
+                            ));
+                  },
+                  changeAvailability: (e) => e.availabilityOption.fold(
+                      () => {},
                       (a) => a.fold(
-                            (l) => showFlushbarError(
-                                errMessage: "Can not update data")
-                              ..show(context),
+                            (l) => Fluttertoast.showToast(
+                                msg: "Availability Not Changed"),
                             (r) {
-                              showFlushbarSuccess(
-                                  succMessage: "Data Updated Successfull")
-                                ..show(context);
-                              r.map(
-                                  loginRequestData: (e) {},
-                                  loginResponseData: (resp) {
-                                    authProvider.setUserData(resp.user);
-                                    context.bloc<AuthenticationBloc>().add(
-                                        AuthenticationEvent.saveAuthentication(
-                                            r: resp.user));
-                                  });
+                              Fluttertoast.showToast(
+                                  msg: "Availability Changed");
+                              final _current = _auth.getUserData().availability;
+                              final _updated = _auth
+                                  .getUserData()
+                                  .copyWith(availability: !_current);
+                              _auth.setUserData(_updated);
                             },
-                          ));
-                },
-                changeAvailability: (e) => e.availabilityOption.fold(
-                    () => {},
-                    (a) => a.fold(
-                          (l) => Fluttertoast.showToast(
-                              msg: "Availability Not Changed"),
-                          (r) {
-                            Fluttertoast.showToast(msg: "Availability Changed");
-                            final _current =
-                                authProvider.getUserData.availability;
-                            final _updated = authProvider.getUserData
-                                .copyWith(availability: !_current);
-                            authProvider.setUserData(_updated);
-                          },
-                        )));
-          },
-          child:
-              BlocConsumer<OtherBloc, OtherState>(listener: (context, state) {
-            state.maybeMap(
-                orElse: () {},
-                getProvinceOption: (e) {
-                  e.getProvinceOption.fold(
-                      () => () {},
-                      (a) => a.fold((l) => print(l), (r) {
-                            setState(() {
-                              province = r;
-                              if (isFirstChanged == false) {
-                                _setProvinceArea(authProvider.getUserData);
-                                context.bloc<OtherBloc>().add(
-                                    OtherEvent.getCityOnly(
-                                        provinceId:
-                                            selectedProvince.provinceId));
-                              }
-                            });
-                          }));
-                },
-                getCityOption: (e) {
-                  e.getCityOption.fold(
-                      () => () {},
-                      (a) => a.fold(
-                            (l) => print(l.toString()),
-                            (r) {
-                              setState(() {
-                                city = r;
-                                if (isFirstChanged == false) {
-                                  _setCityArea(authProvider.getUserData);
-                                }
-                              });
-                            },
-                          ));
-                });
-          }, builder: (context, state) {
-            return GestureDetector(
+                          )));
+            },
+            child: GestureDetector(
               onTap: () {
                 FocusScope.of(context).requestFocus(new FocusNode());
               },
               child: ListView(
                 children: <Widget>[
                   SwitchListTile(
-                    value: authProvider.getUserData.availability,
+                    value: _auth.getUserData().availability,
                     onChanged: (val) {
-                      context.bloc<AuthBloc>().add(
+                      context.read<AuthBloc>().add(
                           AuthEvent.changeGroomerAvailability(
-                              id: authProvider.getUserData.id,
-                              isAvailable:
-                                  !authProvider.getUserData.availability));
+                              id: _auth.getUserData().id,
+                              isAvailable: !_auth.getUserData().availability));
                     },
                     title: Text("Is Available"),
                     subtitle: Text("Is Groomer available ?"),
@@ -334,40 +303,74 @@ class _AdvancedInformationContentState
                                       style: TextStyle(
                                         color: Colors.blue,
                                       )),
-                                  DropdownButton(
-                                    underline: Divider(thickness: 2, height: 2),
-                                    icon: state.maybeMap(
-                                        orElse: () =>
-                                            Icon(Icons.arrow_drop_down),
-                                        getProvinceOption: (e) {
-                                          if (e.isLoading) {
-                                            return CircularProgressIndicator();
-                                          } else {
-                                            return Icon(Icons.arrow_drop_down);
-                                          }
-                                        }),
-                                    hint: Text("Your Province",
-                                        style: TextStyle(
-                                            fontWeight: FontWeight.bold)),
-                                    isExpanded: true,
-                                    items: province
-                                        .map((f) => DropdownMenuItem(
-                                            child: Text(f.province), value: f))
-                                        .toList(),
-                                    value: selectedProvince,
-                                    onChanged: (RajaOngkir val) {
-                                      FocusScope.of(context)
-                                          .requestFocus(FocusNode());
-                                      setState(() {
-                                        isFirstChanged = true;
-                                        selectedProvince = val;
-                                        selectedCity = null;
-                                        context.bloc<OtherBloc>().add(
-                                            OtherEvent.getCityOnly(
-                                                provinceId: val.provinceId));
-                                      });
-                                    },
-                                  ),
+                                  BlocProvider(
+                                    create: (context) => provinceBloc
+                                      ..add(OtherEvent.getProvinceOnly()),
+                                    child: BlocConsumer<OtherBloc, OtherState>(
+                                      listener: (context, state) {
+                                        state.maybeMap(
+                                            orElse: () {},
+                                            getProvinceOption: (e) {
+                                              e.getProvinceOption.fold(
+                                                  () => () {},
+                                                  (a) => a.fold(
+                                                        (l) => print(l),
+                                                        (r) {
+                                                          setState(() {
+                                                            province = r;
+                                                            if (isFirstChanged ==
+                                                                false) {
+                                                              _setProvinceArea(_auth
+                                                                  .getUserData());
+                                                            }
+                                                          });
+                                                        },
+                                                      ));
+                                            });
+                                      },
+                                      builder: (context, state) {
+                                        return DropdownButton(
+                                          underline:
+                                              Divider(thickness: 2, height: 2),
+                                          icon: state.maybeMap(
+                                              orElse: () =>
+                                                  Icon(Icons.arrow_drop_down),
+                                              getProvinceOption: (e) {
+                                                if (e.isLoading) {
+                                                  return CircularProgressIndicator();
+                                                } else {
+                                                  return Icon(
+                                                      Icons.arrow_drop_down);
+                                                }
+                                              }),
+                                          hint: Text("Your Province",
+                                              style: TextStyle(
+                                                  fontWeight: FontWeight.bold)),
+                                          isExpanded: true,
+                                          items: province
+                                              .map((f) => DropdownMenuItem(
+                                                  child: Text(f.province),
+                                                  value: f))
+                                              .toList(),
+                                          value: selectedProvince,
+                                          onChanged: (RajaOngkir val) {
+                                            FocusScope.of(context)
+                                                .requestFocus(FocusNode());
+                                            setState(() {
+                                              isFirstChanged = true;
+                                              selectedProvince = val;
+                                              selectedCity = null;
+
+                                              cityBloc.add(
+                                                  OtherEvent.getCityOnly(
+                                                      provinceId:
+                                                          val.provinceId));
+                                            });
+                                          },
+                                        );
+                                      },
+                                    ),
+                                  )
                                 ],
                               ),
                               SizedBox(height: 10),
@@ -378,36 +381,68 @@ class _AdvancedInformationContentState
                                       style: TextStyle(
                                         color: Colors.blue,
                                       )),
-                                  DropdownButton(
-                                    icon: state.maybeMap(
-                                        orElse: () =>
-                                            Icon(Icons.arrow_drop_down),
-                                        getProvinceOption: (e) {
-                                          if (e.isLoading) {
-                                            return CircularProgressIndicator();
-                                          } else {
-                                            return Icon(Icons.arrow_drop_down);
-                                          }
-                                        }),
-                                    underline: Divider(thickness: 2, height: 2),
-                                    hint: Text("Your City",
-                                        style: TextStyle(
-                                            fontWeight: FontWeight.bold)),
-                                    isExpanded: true,
-                                    items: city
-                                        .map((f) => DropdownMenuItem(
-                                            child: Text(f.cityName.toString()),
-                                            value: f))
-                                        .toList(),
-                                    value: selectedCity,
-                                    onChanged: (RajaOngkir val) {
-                                      FocusScope.of(context)
-                                          .requestFocus(FocusNode());
-                                      setState(() {
-                                        isFirstChanged = true;
-                                        selectedCity = val;
-                                      });
-                                    },
+                                  BlocProvider(
+                                    create: (context) => cityBloc,
+                                    child: BlocConsumer<OtherBloc, OtherState>(
+                                      listener: (context, state) {
+                                        state.maybeMap(
+                                            orElse: () {},
+                                            getCityOption: (e) {
+                                              e.getCityOption.fold(
+                                                  () => () {},
+                                                  (a) => a.fold(
+                                                        (l) =>
+                                                            print(l.toString()),
+                                                        (r) {
+                                                          setState(() {
+                                                            city = r;
+                                                            if (isFirstChanged ==
+                                                                false) {
+                                                              _setCityArea(_auth
+                                                                  .getUserData());
+                                                            }
+                                                          });
+                                                        },
+                                                      ));
+                                            });
+                                      },
+                                      builder: (context, state) {
+                                        return DropdownButton(
+                                          icon: state.maybeMap(
+                                              orElse: () =>
+                                                  Icon(Icons.arrow_drop_down),
+                                              getCityOption: (e) {
+                                                if (e.isLoading) {
+                                                  return CircularProgressIndicator();
+                                                } else {
+                                                  return Icon(
+                                                      Icons.arrow_drop_down);
+                                                }
+                                              }),
+                                          underline:
+                                              Divider(thickness: 2, height: 2),
+                                          hint: Text("Your City",
+                                              style: TextStyle(
+                                                  fontWeight: FontWeight.bold)),
+                                          isExpanded: true,
+                                          items: city
+                                              .map((f) => DropdownMenuItem(
+                                                  child: Text(
+                                                      f.cityName.toString()),
+                                                  value: f))
+                                              .toList(),
+                                          value: selectedCity,
+                                          onChanged: (RajaOngkir val) {
+                                            FocusScope.of(context)
+                                                .requestFocus(FocusNode());
+                                            setState(() {
+                                              isFirstChanged = true;
+                                              selectedCity = val;
+                                            });
+                                          },
+                                        );
+                                      },
+                                    ),
                                   ),
                                 ],
                               ),
@@ -519,33 +554,30 @@ class _AdvancedInformationContentState
                     child: BlocBuilder<AuthBloc, AuthState>(
                         builder: (context, state) {
                       return state.maybeMap(
-                        orElse: () => btnSaveGroomer(authProvider, context),
+                        orElse: () => btnSaveGroomer(_auth, context),
                         onProgress: (e) => BtnPrimaryBlueLoading(),
                         failOrSuccessUpdateCustomerOption: (value) =>
                             value.updateCustomerOption.fold(
                                 () => BtnPrimaryBlueLoading(),
-                                (a) => btnSaveGroomer(authProvider, context)),
+                                (a) => btnSaveGroomer(_auth, context)),
                       );
                     }),
                   ),
                 ],
               ),
-            );
-          }),
-        ),
+            )),
       ),
     );
   }
 
-  BtnPrimaryBlue btnSaveGroomer(
-      AuthProvider authProvider, BuildContext context) {
+  BtnPrimaryBlue btnSaveGroomer(AuthController auth, BuildContext context) {
     return BtnPrimaryBlue(
         text: "Save",
         onPressed: () {
           final _coverageArea =
               "${selectedProvince.province},${selectedCity.cityName},${selectedProvince.provinceId},${selectedCity.cityId}";
 
-          final _updated = authProvider.getUserData.copyWith(
+          final _updated = auth.getUserData().copyWith(
               styling: _style,
               clipping: _clip,
               yearsOfExperience: (_expYearsController.text == null)
@@ -559,7 +591,7 @@ class _AdvancedInformationContentState
               trainingCourses: _courseController.text,
               coverageArea: _coverageArea,
               accessKey: "d78c1a5c-ccbe-4c26-ac08-43ed66c8afb9");
-          context.bloc<AuthBloc>().add(AuthEvent.updateGroomer(user: _updated));
+          context.read<AuthBloc>().add(AuthEvent.updateGroomer(user: _updated));
         });
   }
 
